небольшой рефакторинг для линтера
