+++ public/Components/HomeView/HomeView.js
@@ -44,4 +44,3 @@ export default class HomeView {
     }
 
 }
-
