+++ public/index.html
@@ -20,7 +20,7 @@
   <script src="./modules/inputDataValidation.js"></script>
   <script src="./modules/createElements.js"></script>
   <script src="./modules/actionScript.js"></script>
-  <script src="./modules/ajaxRequests.js" type="module"></script>
+  <script src="./js/ajaxRequests.js" type="module"></script>
   <script src="./js/main.js" type="module"></script>
