--- public/js/script.js
@@ -1,4 +1,0 @@
-setTimeout(() => {
-  container.classList.add('sign-in');
-}, 200);
-
